requirements.txt
python-telegram-bot[job-queue]>=20.0
python-dotenv
requests==2.28.1